Move code shared between Elegant and Bmad conversions
