+++ cheetah/converters/utils/__init__.py
@@ -1,0 +1,1 @@
+from . import fortran_namelist  # noqa: F401
